Update Program.cs (#236)
Set three new license options to true: UseRiskInsights, UseOrganizationDomains, and UseAdminSponsoredFamilies.

As mentioned in #229, licenses had to be regenerated starting in 2025.5.0. It looks like new options were added which default to false unless we set them to true here. I am not sure if the version needs to be bumped or not. I tested locally and I can now access the Claimed Domains page in my organization.

Signed-off-by: Joseph Gigantino <128943406+Jgigantino31@users.noreply.github.com>