Test generating user and organization licenses during build check (#252)
Add commands to build check to build licensegen image as well and test if the created licensegen image can actually generate user and organization licenses. run.sh will print the generated license to stdout and return zero if successful. If an error occurs, a non zero error code is returned which should cause a build error.

Signed-off-by: Joseph Gigantino <128943406+Jgigantino31@users.noreply.github.com>